+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 All notable changes to this project are documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased](https://github.com/orca-services/cakephp-heartbeat/compare/0.1.1...master)
+## [Unreleased](https://github.com/orca-services/cakephp-heartbeat/compare/2.0.0...cakephp-3.x)
+### Added
+- Added documentation on how to write a sensor
+
+### Changed
+
+### Fixed
+
+## [2.0.0](https://github.com/orca-services/cakephp-heartbeat/releases/tag/2.0.0) - 2018-10-01
 ### Added
 - Support for CakePHP Version 3.x
 - Add CakePHP Migrations plugin as development dependency